refactor useExternalLink logic
