+++ src/components/MediaDisplay.tsx
@@ -1,13 +1,12 @@
-import { useState, useEffect, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import Auth from '../auth/Auth';
 import { useGlobal } from 'reactn';
 import { useFetchMediaUrl, MediaSt } from '../crud';
-import { isElectron } from '../api-variable';
-import { launch } from '../utils';
 import { useSnackBar } from '../hoc/SnackBar';
 import { ISharedStrings } from '../model';
 import { sharedSelector } from '../selector';
 import { shallowEqual, useSelector } from 'react-redux';
+import { ExternalLink, LaunchFn } from './ExternalLink';
 
 interface IProps {
   auth: Auth | null;
@@ -17,20 +16,14 @@ interface IProps {
 
 export function MediaDisplay(props: IProps) {
   const { srcMediaId, finish } = props;
-  const [isOffline] = useGlobal('offline');
   const [reporter] = useGlobal('errorReporter');
   const { fetchMediaUrl, mediaState } = useFetchMediaUrl(reporter);
-  const externalRef = useRef<any>();
-  const [externalUrl, setExternalUrl] = useState<string>('#');
+  const launch = useRef<LaunchFn>();
   const { showMessage } = useSnackBar();
   const ts: ISharedStrings = useSelector(sharedSelector, shallowEqual);
 
-  const handleLaunch = (target: string) => {
-    if (isElectron) {
-      launch(target, !isOffline);
-    } else {
-      setExternalUrl(target);
-    }
+  const register = (fn?: LaunchFn) => {
+    launch.current = fn;
   };
 
   useEffect(() => {
@@ -42,8 +35,9 @@ export function MediaDisplay(props: IProps) {
 
   useEffect(() => {
     if (!srcMediaId || mediaState.id !== srcMediaId) return;
-    if (mediaState.status === MediaSt.FETCHED) handleLaunch(mediaState.url);
-    else if (mediaState.error) {
+    if (mediaState.status === MediaSt.FETCHED) {
+      if (launch.current) launch.current(mediaState.url);
+    } else if (mediaState.error) {
       if (mediaState.error.startsWith('no offline file'))
         showMessage(ts.fileNotFound);
       else showMessage(mediaState.error);
@@ -51,23 +45,6 @@ export function MediaDisplay(props: IProps) {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [mediaState]);
 
-  useEffect(() => {
-    if (externalRef.current && externalUrl !== '#') {
-      externalRef.current.click();
-      setExternalUrl('#');
-      finish();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [externalUrl]);
-
-  return (
-    // eslint-disable-next-line jsx-a11y/anchor-has-content
-    <a
-      ref={externalRef}
-      href={externalUrl}
-      target="_blank"
-      rel="noopener noreferrer"
-    ></a>
-  );
+  return <ExternalLink register={register} finish={finish} />;
 }
 export default MediaDisplay;
